reorganize record mapping, add more standard.site options
