+++ packages/links/src/appview.ts
@@ -1,15 +1,21 @@
-import type { AppBskyFeedPost } from "@atproto/api";
 import { uuidv7 } from "uuidv7-js";
-import { type Link, type LinkPost, postType } from "@sill/schema";
-import { serializeBlueskyPostToHtml } from "./bluesky.js";
+import type { Link } from "@sill/schema";
+import { isEmptyRecord, toDbDate, toIso } from "./record-mappers/shared.js";
 
 /**
  * Client for the Sill AppView API — a read-only HTTP index over atproto that
  * answers "what URLs are the people I follow sharing?". See API.md for the
  * full contract. This module wraps the REST endpoints we use for the Bluesky
- * timeline and maps their shapes onto Sill's `link` / `linkPostDenormalized`
- * shapes so the existing frontend renders without changes.
+ * timeline. The per-collection `ShareRow` → `linkPostDenormalized` mappers live
+ * in `./record-mappers/` (dispatched by `shareRowToLinkPost`, re-exported
+ * below); this file keeps the HTTP client, response types, Slingshot
+ * enrichment, and the `UrlItem` → `link` mapping.
  */
+
+// Re-exported so existing `./appview.js` importers keep working after the
+// mapper extraction.
+export { shareRowToLinkPost } from "./record-mappers/index.js";
+export { toDbDate, toIso };
 
 const APPVIEW_TIMEOUT_MS = 60000; // server allows up to 60s during backfill
 
@@ -693,145 +699,7 @@ const fetchRecordFromSlingshot = async (
   }
 };
 
-// --- Mapping helpers ---
-
-/**
- * Parse an AppView/atproto timestamp to an ISO string. Handles both
- * ClickHouse format ("2026-05-20 13:00:00.000", space separator, no Z) and
- * standard ISO record timestamps.
- */
-export const toIso = (ts?: string | null): string | null => {
-  if (!ts) return null;
-  const normalized = ts.includes("T") ? ts : `${ts.replace(" ", "T")}Z`;
-  const d = new Date(normalized);
-  return Number.isNaN(d.getTime()) ? null : d.toISOString();
-};
-
-/**
- * Format a timestamp the way Drizzle's `timestamp({ mode: "string" })` columns
- * return it: UTC, space-separated, no `T`/`Z` ("2026-05-20 13:01:22.500"). The
- * frontend (PostAuthor) appends its own `Z`, so post/quoted/published dates must
- * be in this shape — a full ISO string would produce an invalid date there.
- * Accepts both ClickHouse (AppView) and ISO (record) inputs.
- */
-export const toDbDate = (ts?: string | null): string | null => {
-  const iso = toIso(ts);
-  return iso ? iso.replace("T", " ").replace("Z", "") : null;
-};
-
-const profileUrl = (handleOrDid: string): string =>
-  `https://bsky.app/profile/${handleOrDid}`;
-
-/** Build a bsky.app post permalink from an at:// URI. */
-const postUrlFromAtUri = (atUri: string, handle?: string | null): string => {
-  const parts = atUri.replace("at://", "").split("/");
-  const did = parts[0];
-  const rkey = parts[parts.length - 1];
-  return `https://bsky.app/profile/${handle || did}/post/${rkey}`;
-};
-
-/**
- * Mastodon actor identifier (the ActivityPub Actor URI Sill pushed up — e.g.
- * `https://mastodon.social/users/alice`) is itself a fine profile URL. We
- * leave it untransformed; the federated instance redirects `/users/x` to the
- * human `/@x` view on the client.
- */
-const mastodonProfileUrl = (actorId: string): string => actorId;
-
-/**
- * Heuristic: an `at://…` URI is Bluesky, anything else (http(s)://) is
- * Mastodon. Used to type quoted subjects since `SubjectPost` has no
- * `collection` field — the parent share's collection is the only authoritative
- * signal, and only when the quoted post is reachable.
- */
-const networkFromAtUri = (atUri: string): "bluesky" | "mastodon" =>
-  atUri.startsWith("at://") ? "bluesky" : "mastodon";
-
-/** Quoted-post permalink that respects the subject's network shape. */
-const quotedPostPermalink = (
-  atUri: string,
-  handle: string | null | undefined,
-): string =>
-  networkFromAtUri(atUri) === "bluesky"
-    ? postUrlFromAtUri(atUri, handle)
-    : atUri;
-
-/** Profile URL builder that respects the subject's network shape. */
-const subjectProfileUrl = (
-  atUri: string,
-  handleOrDid: string,
-): string =>
-  networkFromAtUri(atUri) === "bluesky"
-    ? profileUrl(handleOrDid)
-    : mastodonProfileUrl(handleOrDid);
-
-/**
- * Extract attached images from a raw (record-level) atproto post embed,
- * constructing CDN thumbnail URLs from blob refs. Best-effort: returns [] for
- * shapes we don't recognize.
- */
-const extractImagesFromRecord = (
-  // biome-ignore lint/suspicious/noExplicitAny: raw atproto record JSON
-  record: any,
-  did: string
-): { url: string; alt: string }[] => {
-  const embed = record?.embed;
-  if (!embed) return [];
-  // app.bsky.embed.recordWithMedia nests the media under .media
-  const media =
-    embed.$type === "app.bsky.embed.recordWithMedia" ? embed.media : embed;
-  if (
-    media?.$type !== "app.bsky.embed.images" ||
-    !Array.isArray(media.images)
-  ) {
-    return [];
-  }
-  return (
-    media.images
-      // biome-ignore lint/suspicious/noExplicitAny: raw atproto image JSON
-      .map((img: any) => {
-        const cid = img?.image?.ref?.$link ?? img?.image?.ref;
-        if (typeof cid !== "string") return null;
-        return {
-          url: `https://cdn.bsky.app/img/feed_thumbnail/plain/${did}/${cid}@jpeg`,
-          alt: typeof img?.alt === "string" ? img.alt : "",
-        };
-      })
-      .filter((i: { url: string; alt: string } | null) => i !== null)
-  );
-};
-
-const parseRecord = (raw: string): AppBskyFeedPost.Record | null => {
-  try {
-    return JSON.parse(raw) as AppBskyFeedPost.Record;
-  } catch {
-    return null;
-  }
-};
-
-/**
- * The AppView returns unresolved records as an empty object (`"{}"`) rather than
- * omitting them, so a present-but-empty record/subject must be treated as
- * "missing" — otherwise it bypasses the repost fallback and renders blank.
- */
-const isEmptyRecord = (raw?: string | null): boolean => {
-  if (!raw) return true;
-  try {
-    const obj = JSON.parse(raw);
-    return (
-      !obj || typeof obj !== "object" || Object.keys(obj as object).length === 0
-    );
-  } catch {
-    return true;
-  }
-};
-
-const serializeRecord = (record: AppBskyFeedPost.Record | null): string => {
-  // Empty-text posts (image/link-only) may omit `text` entirely; the shared
-  // serializer assumes a string, so guard before calling it.
-  if (!record || typeof record.text !== "string") return "";
-  return serializeBlueskyPostToHtml(record);
-};
+// --- UrlItem → link mapping ---
 
 /**
  * Map an AppView UrlItem to Sill's `link` shape. When an existing DB `link`
@@ -856,640 +724,6 @@ export const urlItemToLink = (item: UrlItem, dbLink?: Link | null): Link => {
     siteName: item.siteName ?? dbLink?.siteName ?? null,
     topics: dbLink?.topics ?? null,
   };
-};
-
-const emptyDenormalized = (share: ShareRow, userId: string): LinkPost => ({
-  id: uuidv7(),
-  linkUrl: share.url,
-  postUrl: "",
-  postText: "",
-  postDate:
-    toDbDate(share.eventTime) ?? toDbDate(new Date().toISOString()) ?? "",
-  postType: postType.enumValues[0], // "bluesky"
-  postImages: [],
-  actorUrl: profileUrl(share.actorHandle || share.actorDid),
-  actorHandle: share.actorHandle || share.actorDid,
-  actorName: share.actorName ?? null,
-  actorAvatarUrl: share.actorAvatar ?? null,
-  quotedActorUrl: null,
-  quotedActorHandle: null,
-  quotedActorName: null,
-  quotedActorAvatarUrl: null,
-  quotedPostUrl: null,
-  quotedPostText: null,
-  quotedPostDate: null,
-  quotedPostType: null,
-  quotedPostImages: null,
-  repostActorUrl: null,
-  repostActorHandle: null,
-  repostActorName: null,
-  repostActorAvatarUrl: null,
-  userId,
-  listId: null,
-});
-
-/**
- * `network.cosmik.card` is a bookmark on Semble. Render it as a bookmark card
- * ("{actor} bookmarked this URL on Semble") linking to the Semble page for the
- * bookmarked URL taken from the record (the raw, non-canonical URL).
- */
-const cosmikCardToLinkPost = (share: ShareRow, base: LinkPost): LinkPost => {
-  let recordUrl = share.url; // fall back to the canonical URL
-  let createdAt: string | undefined;
-  try {
-    const record = JSON.parse(share.record) as {
-      url?: string;
-      createdAt?: string;
-    };
-    if (typeof record.url === "string") recordUrl = record.url;
-    if (typeof record.createdAt === "string") createdAt = record.createdAt;
-  } catch {
-    // keep the canonical URL / share eventTime
-  }
-  const sembleUrl = `https://semble.so/url?id=${encodeURIComponent(recordUrl)}`;
-  const profileUrl = `https://semble.so/profile/${
-    share.actorHandle || share.actorDid
-  }`;
-  return {
-    ...base,
-    postType: postType.enumValues[2], // "atbookmark"
-    // Per-bookmarker URL so two people bookmarking the same link don't collapse
-    // into one card under `groupBy(postUrl)` (the cosmik record is often `{}`,
-    // so every bookmark would otherwise share the same Semble URL). The link to
-    // the Semble page for the URL itself lives in `postText`.
-    postUrl: profileUrl,
-    // Date by when the bookmark was made (falls back to the share eventTime).
-    postDate: toDbDate(createdAt) ?? base.postDate,
-    postText: `Bookmarked this on <a href="${sembleUrl}">Semble</a>.`,
-    actorUrl: profileUrl,
-  };
-};
-
-/**
- * `community.lexicon.bookmarks.bookmark` is a platform-agnostic bookmark of a
- * URL. We don't name or link a platform — the bookmarked URL is the share's URL
- * (shown by the link card) and the bookmarker is the share's actor.
- */
-const communityBookmarkToLinkPost = (
-  share: ShareRow,
-  base: LinkPost
-): LinkPost => {
-  const profile = profileUrl(share.actorHandle || share.actorDid);
-  let createdAt: string | undefined;
-  try {
-    const record = JSON.parse(share.record) as { createdAt?: string };
-    if (typeof record.createdAt === "string") createdAt = record.createdAt;
-  } catch {
-    // fall back to the share eventTime in base
-  }
-  return {
-    ...base,
-    postType: postType.enumValues[2], // "atbookmark"
-    // Per-bookmarker URL so two people bookmarking the same link don't collapse
-    // into one card under `groupBy(postUrl)` (there's no per-bookmark permalink).
-    postUrl: profile,
-    // Date by when the bookmark was made (falls back to the share eventTime).
-    postDate: toDbDate(createdAt) ?? base.postDate,
-    // Sill itself writes these bookmarks; the `sill-bookmark` class flags the
-    // card for the Sill logo in PostRep.
-    postText: '<span class="sill-bookmark">Bookmarked this URL.</span>',
-    actorUrl: profile,
-  };
-};
-
-/**
- * `app.rocksky.scrobble` is a music listen logged by Rocksky. The AppView
- * aggregates it under the track's streaming URL (the `spotifyLink`), so the
- * link card already shows the track + album art; this builds a one-line
- * "Listened to {title} by {artist}" post body with an album/year subline. The
- * `scrobble-card` marker class flags it for the Rocksky logo in PostRep.
- *
- * Styles are inline rather than via PostContent.module.css — the body is
- * injected through `dangerouslySetInnerHTML`, so CSS-module-scoped class names
- * wouldn't match.
- */
-interface ScrobbleRecord {
-  title?: string;
-  artist?: string;
-  album?: string;
-  albumArtist?: string;
-  year?: number;
-  genre?: string;
-  createdAt?: string;
-  spotifyLink?: string;
-  tidalLink?: string;
-  appleMusicLink?: string;
-  youtubeLink?: string;
-}
-
-const scrobbleToLinkPost = (share: ShareRow, base: LinkPost): LinkPost => {
-  let record: ScrobbleRecord = {};
-  try {
-    record = JSON.parse(share.record) as ScrobbleRecord;
-  } catch {
-    // leave record empty → fall back to the canonical URL / share eventTime
-  }
-
-  const title = record.title?.trim() || "a track";
-  const artist = record.artist?.trim() || record.albumArtist?.trim() || "";
-  const album = record.album?.trim() || "";
-  // Link the track to the best available streaming URL; the aggregated share
-  // URL (usually the Spotify link) is the safe default.
-  const trackHref =
-    safeHref(share.url) ||
-    safeHref(record.spotifyLink ?? "") ||
-    safeHref(record.tidalLink ?? "") ||
-    safeHref(record.appleMusicLink ?? "") ||
-    safeHref(record.youtubeLink ?? "");
-  const profile = profileUrl(share.actorHandle || share.actorDid);
-
-  // Secondary line: "Album · 2024 · Genre" (only the parts we have).
-  const subParts = [
-    album,
-    typeof record.year === "number" ? String(record.year) : "",
-    record.genre?.trim() ?? "",
-  ].filter(Boolean);
-
-  const titleHtml = trackHref
-    ? `<a href="${escapeAttr(trackHref)}"><strong>${escapeHtml(title)}</strong></a>`
-    : `<strong>${escapeHtml(title)}</strong>`;
-  const line = artist
-    ? `Listened to ${titleHtml} by ${escapeHtml(artist)}`
-    : `Listened to ${titleHtml}`;
-
-  const postText =
-    `<div class="scrobble-card">` +
-    `<div>${line}</div>` +
-    (subParts.length
-      ? `<div style="font-size: var(--font-size-1); color: var(--gray-11); margin-top: var(--space-1)">${escapeHtml(
-          subParts.join(" · "),
-        )}</div>`
-      : "") +
-    `</div>`;
-
-  return {
-    ...base,
-    postType: postType.enumValues[2], // "atbookmark" (non-bsky/mastodon collection)
-    // Per-scrobbler URL so repeat listens by one person collapse into one card
-    // while different listeners stay distinct (mirrors the bookmark handling).
-    postUrl: profile,
-    postDate: toDbDate(record.createdAt) ?? base.postDate,
-    postText,
-    actorUrl: profile,
-  };
-};
-
-// --- site.standard.document (leaflet content) ---
-
-// A `site.standard.document` is a long-form post (a blog entry). The AppView
-// indexes it as a "share" of any URL linked inside its body. We render the one
-// paragraph that links to the shared URL and point the card at the original
-// document (the publication's base URL + the document's path). Content can be
-// in several lexicons; `pub.leaflet.content` is the only one handled so far.
-
-const escapeHtml = (s: string): string =>
-  s
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;");
-
-const escapeAttr = (s: string): string => escapeHtml(s).replace(/"/g, "&quot;");
-
-/** Only allow http(s) hrefs through (no `javascript:` etc.). */
-const safeHref = (uri: string): string | null =>
-  /^https?:\/\//i.test(uri) ? uri : null;
-
-interface LeafletFacet {
-  index?: { byteStart: number; byteEnd: number };
-  features?: { $type?: string; uri?: string }[];
-}
-
-const LEAFLET_FACET = "pub.leaflet.richtext.facet";
-
-/** Wrap a (already HTML-escaped) text run in the markup for its facet features. */
-const wrapLeafletFeatures = (
-  text: string,
-  features: { $type?: string; uri?: string }[]
-): string => {
-  let html = text;
-  let href: string | null = null;
-  for (const f of features) {
-    switch (f.$type) {
-      case `${LEAFLET_FACET}#link`:
-        if (typeof f.uri === "string") href = safeHref(f.uri);
-        break;
-      case `${LEAFLET_FACET}#bold`:
-        html = `<strong>${html}</strong>`;
-        break;
-      case `${LEAFLET_FACET}#italic`:
-        html = `<em>${html}</em>`;
-        break;
-      case `${LEAFLET_FACET}#code`:
-        html = `<code>${html}</code>`;
-        break;
-      case `${LEAFLET_FACET}#strikethrough`:
-        html = `<s>${html}</s>`;
-        break;
-      case `${LEAFLET_FACET}#underline`:
-        html = `<u>${html}</u>`;
-        break;
-      // highlight/footnote/mentions: rendered as plain text for now.
-    }
-  }
-  if (href) html = `<a href="${escapeAttr(href)}">${html}</a>`;
-  return html;
-};
-
-/**
- * Render a leaflet rich-text block (plaintext + `pub.leaflet.richtext.facet[]`)
- * to HTML. Facet indices are byte offsets into the UTF-8 text, so we slice a
- * byte view and decode each run (a JS string slice would mis-index on any
- * multi-byte character).
- */
-const renderLeafletRichText = (
-  plaintext: string,
-  facets?: LeafletFacet[]
-): string => {
-  const bytes = new TextEncoder().encode(plaintext);
-  const decoder = new TextDecoder();
-  const slice = (start: number, end: number): string =>
-    escapeHtml(decoder.decode(bytes.slice(start, end))).replace(
-      /\n/g,
-      "<br />"
-    );
-
-  const valid = (facets ?? [])
-    .filter(
-      (f): f is Required<Pick<LeafletFacet, "index">> & LeafletFacet =>
-        !!f.index &&
-        Array.isArray(f.features) &&
-        f.index.byteEnd > f.index.byteStart
-    )
-    .sort((a, b) => a.index.byteStart - b.index.byteStart);
-
-  const out: string[] = [];
-  let cursor = 0;
-  for (const f of valid) {
-    const { byteStart, byteEnd } = f.index;
-    if (byteStart < cursor) continue; // skip overlapping facets
-    if (byteStart > cursor) out.push(slice(cursor, byteStart));
-    out.push(wrapLeafletFeatures(slice(byteStart, byteEnd), f.features ?? []));
-    cursor = byteEnd;
-  }
-  if (cursor < bytes.length) out.push(slice(cursor, bytes.length));
-  return out.join("");
-};
-
-const stripTrailingSlash = (u: string): string => u.replace(/\/+$/, "");
-
-/** Trailing-slash- and query-insensitive URL match, for locating a link facet. */
-const urlsMatchForFacet = (a: string, b: string): boolean => {
-  const na = stripTrailingSlash(a).toLowerCase();
-  const nb = stripTrailingSlash(b).toLowerCase();
-  if (na === nb) return true;
-  const bare = (u: string): string => {
-    try {
-      const x = new URL(u);
-      x.hash = "";
-      x.search = "";
-      return stripTrailingSlash(x.toString()).toLowerCase();
-    } catch {
-      return stripTrailingSlash(u).toLowerCase();
-    }
-  };
-  return bare(a) === bare(b);
-};
-
-// biome-ignore lint/suspicious/noExplicitAny: raw leaflet record JSON
-type LeafletRecord = any;
-
-/**
- * Find the first leaflet block (text/header/blockquote — anything with
- * `plaintext` + `facets`) whose facets link to `targetUrl`, and return it
- * rendered to HTML. Handles both `linearDocument` and `canvas` pages (both
- * carry blocks as `{ block: <inner> }`). Returns null when the content isn't
- * `pub.leaflet.content` or no block links to the URL.
- */
-const findLeafletLinkParagraph = (
-  record: LeafletRecord,
-  targetUrl: string
-): string | null => {
-  const content = record?.content;
-  if (!content || content.$type !== "pub.leaflet.content") return null;
-  const pages = Array.isArray(content.pages) ? content.pages : [];
-  for (const page of pages) {
-    const blocks = Array.isArray(page?.blocks) ? page.blocks : [];
-    for (const wrapper of blocks) {
-      const inner = wrapper?.block;
-      if (
-        !inner ||
-        typeof inner.plaintext !== "string" ||
-        !Array.isArray(inner.facets)
-      ) {
-        continue;
-      }
-      const links = inner.facets.some((f: LeafletFacet) =>
-        f?.features?.some(
-          (ft) =>
-            ft?.$type === `${LEAFLET_FACET}#link` &&
-            typeof ft.uri === "string" &&
-            urlsMatchForFacet(ft.uri, targetUrl)
-        )
-      );
-      if (links) return renderLeafletRichText(inner.plaintext, inner.facets);
-    }
-  }
-  return null;
-};
-
-/**
- * Build the public document URL: the publication's base URL (resolved by
- * `resolveLeafletPublications`, or an https `site` used directly) joined with
- * the document's `path`. Returns null when no base URL is known.
- */
-const buildLeafletDocUrl = (
-  record: LeafletRecord,
-  publicationUrl?: string
-): string | null => {
-  let base = publicationUrl;
-  if (!base) {
-    const site = typeof record?.site === "string" ? record.site : "";
-    if (/^https?:\/\//i.test(site)) base = site;
-  }
-  if (!base) return null;
-  const b = stripTrailingSlash(base);
-  const path = typeof record?.path === "string" ? record.path : "";
-  if (!path) return b;
-  return `${b}${path.startsWith("/") ? path : `/${path}`}`;
-};
-
-/**
- * Map a `site.standard.document` share to Sill's `linkPostDenormalized` shape:
- * the card body is the paragraph that links to the shared URL (rendered rich
- * text) plus a line linking to the full document on its publication (leaflet).
- */
-const leafletDocumentToLinkPost = (
-  share: ShareRow,
-  base: LinkPost
-): LinkPost => {
-  let record: LeafletRecord = null;
-  try {
-    record = JSON.parse(share.record);
-  } catch {
-    // leave record null → bare fallback below
-  }
-  const docUrl = buildLeafletDocUrl(record, share.publicationUrl);
-  const title =
-    typeof record?.title === "string" && record.title ? record.title : "a post";
-  const docLink = docUrl
-    ? `<a href="${escapeAttr(docUrl)}">${escapeHtml(title)}</a>`
-    : escapeHtml(title);
-
-  const publication = escapeHtml(share.publicationName || "Leaflet");
-
-  const paragraph = record
-    ? findLeafletLinkParagraph(record, share.url)
-    : null;
-  // When the paragraph is found, quote it as a blockquote with a "from {doc}"
-  // line; otherwise a generic line still linking to the document. The
-  // `leaflet-source` class also flags the card for the Leaflet logo in PostRep.
-  const postText = paragraph
-    ? `<blockquote>${paragraph}</blockquote><p class="leaflet-source">From ${docLink} on ${publication}</p>`
-    : `<p class="leaflet-source">Linked to this in ${docLink} on ${publication}.</p>`;
-
-  return {
-    ...base,
-    postType: postType.enumValues[2], // "atbookmark" (non-bsky/mastodon collection)
-    // The card timestamp links to the document. Falls back to the author's
-    // profile only when the document URL can't be resolved.
-    postUrl: docUrl ?? base.actorUrl,
-    // Date the card by when the document was published, not when the AppView
-    // indexed the share (falls back to the share eventTime from `base`).
-    postDate: toDbDate(record?.publishedAt) ?? base.postDate,
-    postText,
-    actorUrl: profileUrl(share.actorHandle || share.actorDid),
-  };
-};
-
-type QuotedFields = Pick<
-  LinkPost,
-  | "quotedActorUrl"
-  | "quotedActorHandle"
-  | "quotedActorName"
-  | "quotedActorAvatarUrl"
-  | "quotedPostUrl"
-  | "quotedPostText"
-  | "quotedPostDate"
-  | "quotedPostType"
-  | "quotedPostImages"
->;
-
-/**
- * Build Sill's `quoted*` fields from a resolved quoted post (a SubjectPost).
- * The subject carries no `collection` of its own, so we infer the network from
- * the at:// vs http(s) shape of `atUri` — Bluesky quotes get a bsky.app
- * permalink and `quotedPostType: "bluesky"`; Mastodon quotes pass the HTTP
- * URL through and get `quotedPostType: "mastodon"`.
- */
-const quotedFields = (subject: SubjectPost): QuotedFields => {
-  const quoted = parseRecord(subject.record);
-  const isBsky = networkFromAtUri(subject.atUri) === "bluesky";
-  const handleOrDid = subject.actorHandle || subject.actorDid;
-  return {
-    quotedActorUrl: subjectProfileUrl(subject.atUri, handleOrDid),
-    quotedActorHandle: handleOrDid,
-    quotedActorName: subject.actorName ?? null,
-    quotedActorAvatarUrl: subject.actorAvatar ?? null,
-    quotedPostUrl: quotedPostPermalink(subject.atUri, subject.actorHandle),
-    quotedPostText: serializeRecord(quoted),
-    quotedPostDate: toDbDate(quoted?.createdAt),
-    quotedPostType: isBsky ? postType.enumValues[0] : postType.enumValues[1], // "bluesky" | "mastodon"
-    quotedPostImages: isBsky
-      ? extractImagesFromRecord(quoted, subject.actorDid)
-      : [],
-  };
-};
-
-/**
- * Map an AppView hydration ShareRow to Sill's `linkPostDenormalized` shape.
- *
- * - `app.bsky.feed.post`: a normal post or a quote. `actor*` is the sharer; a
- *   `subject` (if present) is the quoted post → `quoted*` fields.
- * - `app.bsky.feed.repost`: `actor*` becomes the original author (from
- *   `subject`) and the reposter moves to `repostActor*`, matching Sill's model
- *   where the card body is the original post with a "reposted by" banner. If
- *   the reposted post is itself a quote, its quoted post (`subject.subject`)
- *   fills the `quoted*` fields.
- * - `network.cosmik.card`: a Semble bookmark (see `cosmikCardToLinkPost`).
- * - `community.lexicon.bookmarks.bookmark`: a platform-agnostic bookmark (see
- *   `communityBookmarkToLinkPost`).
- * - `site.standard.document`: a long-form post linking to the URL in its body
- *   (see `leafletDocumentToLinkPost`).
- * - `app.rocksky.scrobble`: a music listen, rendered as a compact track card
- *   (see `scrobbleToLinkPost`).
- *
- * Quoted-post cards and repost original-authors are only available when the
- * AppView resolved `subject` (in-network author); otherwise we fall back to a
- * bare share attributed to the sharer.
- */
-export const shareRowToLinkPost = (
-  share: ShareRow,
-  userId: string
-): LinkPost => {
-  const base = emptyDenormalized(share, userId);
-
-  if (share.collection === "network.cosmik.card") {
-    return cosmikCardToLinkPost(share, base);
-  }
-
-  if (share.collection === "community.lexicon.bookmarks.bookmark") {
-    return communityBookmarkToLinkPost(share, base);
-  }
-
-  if (share.collection === "site.standard.document") {
-    return leafletDocumentToLinkPost(share, base);
-  }
-
-  if (share.collection === "app.rocksky.scrobble") {
-    return scrobbleToLinkPost(share, base);
-  }
-
-  if (share.collection === "mastodon.status") {
-    return mastodonStatusToLinkPost(share, base);
-  }
-
-  if (share.collection === "mastodon.repost") {
-    return mastodonRepostToLinkPost(share, base);
-  }
-
-  if (share.collection === "app.bsky.feed.repost") {
-    // Absent or empty (`{}`) subject → unresolved original; render bare.
-    if (!share.subject || isEmptyRecord(share.subject.record)) return base;
-    const subjectRecord = parseRecord(share.subject.record);
-    // Repost OF a quote post: the quoted post is the subject's own subject.
-    const nestedQuote =
-      share.subject.subject && !isEmptyRecord(share.subject.subject.record)
-        ? quotedFields(share.subject.subject)
-        : null;
-    return {
-      ...base,
-      // original author becomes the primary actor; the card body is the
-      // original post, so its createdAt is the date we render.
-      postUrl: postUrlFromAtUri(share.subject.atUri, share.subject.actorHandle),
-      postText: serializeRecord(subjectRecord),
-      postDate: toDbDate(subjectRecord?.createdAt) ?? base.postDate,
-      postImages: extractImagesFromRecord(
-        subjectRecord,
-        share.subject.actorDid
-      ),
-      actorUrl: profileUrl(share.subject.actorHandle || share.subject.actorDid),
-      actorHandle: share.subject.actorHandle || share.subject.actorDid,
-      actorName: share.subject.actorName ?? null,
-      actorAvatarUrl: share.subject.actorAvatar ?? null,
-      // the network member who reposted
-      repostActorUrl: profileUrl(share.actorHandle || share.actorDid),
-      repostActorHandle: share.actorHandle || share.actorDid,
-      repostActorName: share.actorName ?? null,
-      repostActorAvatarUrl: share.actorAvatar ?? null,
-      ...nestedQuote,
-    };
-  }
-
-  // app.bsky.feed.post (normal post or quote post)
-  const record = parseRecord(share.record);
-  const post: LinkPost = {
-    ...base,
-    postUrl: postUrlFromAtUri(share.atUri, share.actorHandle),
-    postText: serializeRecord(record),
-    postDate: toDbDate(record?.createdAt) ?? base.postDate,
-    postImages: extractImagesFromRecord(record, share.actorDid),
-  };
-
-  if (share.subject && !isEmptyRecord(share.subject.record)) {
-    Object.assign(post, quotedFields(share.subject));
-  }
-
-  return post;
-};
-
-/**
- * `mastodon.status`: a Mastodon post or quote-post. The AppView synthesises
- * the record body with `text`/`createdAt`/`uri`; `atUri` is the Mastodon HTTP
- * permalink (use as-is). `actorDid` is the ActivityPub Actor URI Sill pushed
- * up — its own profile URL. When `subject` is set the post is a quote;
- * `quotedFields` infers Mastodon vs Bluesky from the subject's `atUri` shape.
- */
-const mastodonStatusToLinkPost = (
-  share: ShareRow,
-  base: LinkPost,
-): LinkPost => {
-  const record = parseMastodonRecord(share.record);
-  const post: LinkPost = {
-    ...base,
-    postType: postType.enumValues[1], // "mastodon"
-    postUrl: share.atUri,
-    postText: record?.text ?? "",
-    postDate: toDbDate(record?.createdAt) ?? base.postDate,
-    actorUrl: mastodonProfileUrl(share.actorDid),
-  };
-
-  if (share.subject && !isEmptyRecord(share.subject.record)) {
-    Object.assign(post, quotedFields(share.subject));
-  }
-  return post;
-};
-
-/**
- * `mastodon.repost`: a reblog of a Mastodon status. Same shape convention as
- * `app.bsky.feed.repost` — the reposted post lives in `subject`; the sharer
- * (actor*) is collapsed into `repostActor*` while the original author rides
- * up to the primary actor slot. Mastodon-flavoured URLs throughout.
- */
-const mastodonRepostToLinkPost = (
-  share: ShareRow,
-  base: LinkPost,
-): LinkPost => {
-  // No subject → unresolved (out-of-network original). Render the sharer with
-  // a `mastodon` type so it isn't mis-typed as Bluesky.
-  if (!share.subject || isEmptyRecord(share.subject.record)) {
-    return {
-      ...base,
-      postType: postType.enumValues[1], // "mastodon"
-      actorUrl: mastodonProfileUrl(share.actorDid),
-    };
-  }
-  const subjectRecord = parseMastodonRecord(share.subject.record);
-  const nestedQuote =
-    share.subject.subject && !isEmptyRecord(share.subject.subject.record)
-      ? quotedFields(share.subject.subject)
-      : null;
-  return {
-    ...base,
-    postType: postType.enumValues[1], // "mastodon"
-    postUrl: share.subject.atUri,
-    postText: subjectRecord?.text ?? "",
-    postDate: toDbDate(subjectRecord?.createdAt) ?? base.postDate,
-    actorUrl: mastodonProfileUrl(share.subject.actorDid),
-    actorHandle: share.subject.actorHandle || share.subject.actorDid,
-    actorName: share.subject.actorName ?? null,
-    actorAvatarUrl: share.subject.actorAvatar ?? null,
-    repostActorUrl: mastodonProfileUrl(share.actorDid),
-    repostActorHandle: share.actorHandle || share.actorDid,
-    repostActorName: share.actorName ?? null,
-    repostActorAvatarUrl: share.actorAvatar ?? null,
-    ...nestedQuote,
-  };
-};
-
-/** Minimal parse for the synthesized Mastodon record body (`{text, createdAt, uri}`). */
-const parseMastodonRecord = (
-  raw: string,
-): { text?: string; createdAt?: string; uri?: string } | null => {
-  try {
-    return JSON.parse(raw) as { text?: string; createdAt?: string; uri?: string };
-  } catch {
-    return null;
-  }
 };
 
 /**
